Outline anim-based $
